refactor(domainowned): reorganize components
